(WIP)Implemented CreateDaysForTrip and BottomSheet for TripDetailScreen

+++ app/src/main/java/com/example/trevia/data/schedule/OfflineTripRepository.kt
@@ -15,7 +15,7 @@ class OfflineTripRepository @Inject constructor(
 
     override suspend fun updateTrip(trip: Trip) = tripDao.update(trip)
 
-    override suspend fun deleteTripById(tripId: Int) = tripDao.deleteTripById(tripId)
+    override suspend fun deleteTripById(tripId: Long) = tripDao.deleteTripById(tripId)
 
     override fun getAllTripsStream(): Flow<List<TripModel>>
     {
